Update main.c
Added a function that returns a string example "myFile.txt, myFile.jpg instead of user having to type out the exact path which they will end up creating errors or misplacing the file in their system unable to decrypt the file again. Next change will be focusing on a en/decrypt method that will handle large files such as a mp4... which we can speed the process up a bit by getting creative.